instead of assigning `wImg`, `hImg` in draw, assert they are equal
Asserts that the image width and height are the same in the parent
viewport as they are in the child viewport.

They *should* always be, since we copy the sizes from the parent on
creation with `addViewport`. This should discourage the usage of
`initViewport` unless strictly needed, in which case the caller will
have to make sure that `wImg` and `hImg` are really the same.