create info,warn,debug dummy templates
Currently all echoing is only done if compiled with the `-d:verbose`
flag. This is okayish for now, but there need to be different log
levels. We could think about introducing actual logging, but I think
that's too much. In any case warning e.g. should be echoed always in general.