add `tickLabels`, allows for manual labels at specific locations
Instead of always using the label that corresponds to the number,
`tickLabels` allows to create labels that do not numerically match to
the tick position.

Useful for different scalings, e.g. data in log scale, but labels
showing non log space numbers (or simply for string tick labels).

TODO: think about whether to call updateDataScale here in this proc!