add rudimentary support for `ukStrWidth`
Mainly w/ support to convert strWidth to points / relative so that we
can set certain values depending on their string width.
Arithmetics are not supported with it.

Also to keep in mind that strWidth essentially is rather a Quantity
than a Coord! But for simplicities sake (don'thave to change impl of
Quantity) can support this via Coord.