add WIP auto subtraction for tick labels
If the tick labels of a plot are all such that the string
representation is the same, we now subtract the minimum value from all
tick label texts and add it as a text in the corner.

e.g. if all are represented as 1.342321231e+16, this will subtract
that value and show only the significant digits.