improve formatting of zero tick values as strings
We introduce a heuristic rule to determine if a value is likely to be
zero in a given scale by comparing it with a 1/10 of the difference of
two ticks (or in some cases a 1/100 of the scale of the associated
axis).
For automatically calculated tick positions this should work out to
detect the zero value and thus replace it by the "0" string.

If the detection fails the worst that happens is that the "real"
string representation using 4 significant digits will be used (and we
just made the code for tick labels even more ugly... *sigh*).