allow `ukRelative` coord as arg to `initTickLabel`
When creating a discrete label in ggplotnim it's desirable to be able
to hand the coordinate of that tick position in relative coordinates,
because discrete values are created by splitting the main viewport
into one viewport for each discrete value. Then tick labels at 0.5 in
ukRelative is in the middle.