add `embedAt` which embeds and updates children width/height
`embedAt` is aimed at users of the ginger API and can be used to embed
a given viewport into a new root at a certain child index. This proc
then performs the actual embedding and updates the childrens' widths
and heights.

This is useful to combine multiple finished viewports into a larger viewport.