fix arithmetic for `Coord1D`
For cases where they are not compatible, the operation will still be
attempted in absolute or data units.

Conversion to relative units is now the last resort!

This allows to keep an absolute scale absolute. Previously a
coordinate after embedding twice would result in bad embeddings, since
the absolute coordinate was turned relative and thus subject to bad
conversions (read: not what the user intended using absolute units).