add PContext field to BImage, only create one context
Instead of creating a new context for each cairo call, we now create
one global context on which we perform all actions.

In order to make this work, we save the current context before any
actions and revert it afterwards.

This is not majorly more efficient, since the saving and restoring is
almost as costly as creating a whole new context, but it's still
somewhat better (tested for a scatter plot with 1 Mio. points).

Also the cairo impact then is visible in the final =ctx.destroy= call
in =backends.nim=.