properly handle usage of `sub,..` procs for Coord1D of kind ukData
In the arithmetic procs for Coord1D we use `sub, add, times,
divide`. However, in those we may (for ukData kind) convert the uk
data coord to a relative quantity.

However, if the quantity for which `toRelative` from ukData is called
represents a coordinate, the following line:
rel = q.val / (scale.high - scale.low)
is obviously wrong, since that gives us the quantity `q.val` reresents
on a scale defined by `scale.high - scale.low`.

Thus we have a new `asCoordinate` argument for quantity sub,
... procs, which first subtract the `scale.low` value from `q.val` to
essentially get the correct result for a coordinate.