convert tick label positions to relative
While we want to be able to use absolute units when adding a margin to
a tick label, keeping it as absolute is problematic, because the
position then won't be properly moved when being embedded. For facet
plots in ggplotnim this results in x tick labels, which are not
actually at the plot, but still at the bottom of the plot.