fix margin handling in `layout`
Instead of basing the given margin on the total size of the given
viewport, the size should be relative to each *resulting* viewport. So
that 0.1 relative stays 10% no matter how many elements are created.

It might still be better if we would take possible column widths or
heights into consideration.