extend viewport by FileTypeKind field for TextExtent information
The Cairo backend *needs* the correct file type for the surface to
draw on to get text extents. It's something I completely overlooked
so far, leading to slight inaccuracies of placing text!
One of the multiple reasons our text placement was never quite what I
wanted.

Also, `strWidth / strHeight` now allow to set the text that will be
used, not only the `getStrWidth/Height` procs.