[TikZ] handle number of lines for getStrHeight for TikZ dep on escaping
If input is being escaped and thus `\n` replaced by `\\` we will
get the correct height reported by LaTeX directly. This is despite us
still printing each line in the output 'by hand' as individual lines
of text.

If the input is explicitly `\\` multiline the text is seen as single
line anywany and we write a `tabular` environment and a single text
element on the ginger side.